split out naive unification carrier to separate module
